# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organizations Directory API",
        "description": "API for managing organizations, buildings, and activities",
        "version": "1.0.0"
    },
    "paths": {
        "/api/organizations/by-building/{building_id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organizations in a building",
                "operationId": "31ed562b0669bf019a90d68bf45475bc",
                "parameters": [
                    {
                        "name": "building_id",
                        "in": "path",
                        "description": "Building ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationCollection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Building not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/by-activity/{activity_id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organizations by activity",
                "operationId": "92102a7a5b3033240c1bc30848661c46",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "path",
                        "description": "Activity ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationCollection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Activity not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/nearby": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Find organizations nearby",
                "operationId": "2c4aaebf1645ecdcad5af01f7b635407",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude of center point",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Longitude of center point",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Search radius in meters",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bbox",
                        "in": "query",
                        "description": "Bounding box coordinates (min_lat,min_lng,max_lat,max_lng)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationCollection"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organization by ID",
                "operationId": "68e75650b204083886e125e02aaa2f4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Organization ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationFull"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organization not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/search-by-activity": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Search organizations by activity hierarchy",
                "operationId": "ac7fe46e5aafabf9242c634018ede77d",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "query",
                        "description": "Root activity ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/organizations/search-by-name": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Search organizations by name",
                "operationId": "aa6f8f9ab2a6770c3be4762afc3b1570",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Organization name or part of name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Organization": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Рога и Копыта"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "2-222-222"
                        }
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "OrganizationFull": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Organization"
                    },
                    {
                        "properties": {
                            "building": {
                                "$ref": "#/components/schemas/Building"
                            },
                            "activities": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Building": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Ленина 1, офис 3"
                    },
                    "coordinates": {
                        "properties": {
                            "latitude": {
                                "type": "number",
                                "format": "float",
                                "example": 55.755826
                            },
                            "longitude": {
                                "type": "number",
                                "format": "float",
                                "example": 37.6173
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Activity": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Молочная продукция"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "level": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "OrganizationCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Organization"
                        }
                    },
                    "meta": {
                        "properties": {
                            "count": {
                                "type": "integer",
                                "example": 5
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Organizations",
            "description": "Organizations"
        }
    ]
}